Fix not asking for autoCompile when it's undefined
